fix: JudgeList queue

+++ src/main/java/com/example/codeE/service/judge/JudgeImpl.java
@@ -135,8 +135,8 @@ public class JudgeImpl implements JudgeService {
         updates.setMemory(null);
         updates.setScore(null);
         updates.setResult(null);
-        updates.setCasePoints((float) 0);
-        updates.setCaseTotal((float) 0);
+        updates.setCasePoints(0.0);
+        updates.setCaseTotal(0.0);
         updates.setError(null);
         updates.setDateGrade(rejudge ? DateTimeUtil.format(LocalDateTime.now()) : null);
         updates.setStatus("QU");
